MR comments

+++ data/crac/crac-impl/src/test/java/com/farao_community/farao/data/crac_impl/HvdcRangeImplTest.java
@@ -1,12 +1,9 @@
 /*
- *
- *  * Copyright (c) 2020, RTE (http://www.rte-france.com)
- *  * This Source Code Form is subject to the terms of the Mozilla Public
- *  * License, v. 2.0. If a copy of the MPL was not distributed with this
- *  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- *
+ * Copyright (c) 2021, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-
 package com.farao_community.farao.data.crac_impl;
 
 import com.farao_community.farao.data.crac_api.range_action.RangeType;
@@ -21,8 +18,8 @@ import static org.junit.Assert.assertNotEquals;
  */
 public class HvdcRangeImplTest {
 
-    private final int min = 1;
-    private final int max = 32;
+    private final double min = 1;
+    private final double max = 32;
 
     private HvdcRangeImpl fixedRange;
